Refactor to next.js recommendation
